requirements.txt
amqp==5.2.0
APScheduler==3.10.4
async-timeout==4.0.3
backports.tarfile==1.2.0
bidict==0.23.1
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
docutils==0.21.2
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.2.1
kiwisolver==1.4.5
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
nh3==0.2.18
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pkginfo==1.10.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
pyloco==0.0.139
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
readme_renderer==44.0
redis==5.0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rq==1.16.2
SecretStorage==3.3.3
simple-websocket==1.0.0
SimpleWebSocketServer==0.1.2
six==1.16.0
SQLAlchemy==2.0.31
twine==5.1.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
ushlex==0.99.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
